Add logic to detect the compiler version and the C++ language standard used.
This is largely adapted from the Cloud C++ APIs with the following changes:

* added C++20 to the list of possible standards;
* added "latest" as a possible C++ standard for MSVC;
* added the C++ standard library vendor to the compiler info string, primarily
  to track STLPort.

Here are some sample values I'm getting:
* on my Mac laptop: `AppleClang-11.0.3.11030032-ex-2011-libcpp`;
* google3 Android build: `Clang-9.0.8-ex-2017-libcpp`;
* google3 Windows build: `MSVC-19.22.27905-ex-2017-msvc`;
* google3 Linux build: `Clang-9999.0.0-noex-2017-libcpp` (unfortunately,
  Clang trunk in google3 simply sets the version to 9999:
  https://source.corp.google.com/piper///depot/google3/third_party/llvm/llvm-project/clang/BUILD;l=374-383;rcl=330443088)

Note that as long as the C++ SDKs are distributed in precompiled form, the compiler version reported will be from the compiler we used to compile the binaries, not the compiler used by our users. Unfortunately, there is no clear way to avoid this that wouldn't cause an ODR violation (if e.g. `CompilerId` were placed in a header as an inline function, it would end up having different definitions depending on whether it's included internally or by the user (indirectly)). However, this shouldn't be a big issue as the metrics will still allow tracking total C++ usage and, to an extent, usage by platform.

PiperOrigin-RevId: 331654266