Automated g4 rollback of changelist 326267389.
*** Reason for rollback ***

Roll forward cl/325229850: Improve the tests for ClearPersistence() to use the default Firebase app instead of a custom one.

I've added a workaround for the broken test, a call to Terminate() in Android, and created b/168628900 to track the underlying bug.

*** Original change description ***

Automated g4 rollback of changelist 325229850.

*** Reason for rollback ***

firestore_test_android appears to have been broken by this change (b/163526671).

*** Original change description ***

Improve the tests for ClearPersistence() to use the default Firebase app instead of a custom one.

This fixes a concern raised by @pengzk in cl/309775646 that the Firestore* pointers were leaking. By using the default Firestore object retrieved via firestore() instead of a custom Firestore object return...

***

PiperOrigin-RevId: 331991549