Temporarily add a platform token to the Cloud API headers.
Until b/135633112 is resolved and support for Firebase platform logging is
rolled out on the Firestore backend, there is no good way to estimate the
number of users of the Firestore C++ SDK by platform. As a workaround, add
a platform token to the Cloud headers. Proper support for platform tokens was
actually never added to Cloud headers, but the data processing pipeline
simply ignores unknown tokens so there is no harm in doing this addition.

The implementation reuses the notion of a client language in the Core API to
lump together two tokens (the language and the platform). This is deemed
acceptable only due to the temporary nature of this workaround.

Sample of the header values I'm seeing:
- running C++ tests on Linux in google3:
  `gl-cpp/Clang-9999.0.0-noex-2017-libcpp gl-linux/ fire/1.17.1 grpc/1.33.0-dev`
  (notice the `gl-linux/` token);
- in Unity: `gl-dotnet/4.0.30319 gl-macos/ fire/1.17.1 grpc/1.33.0-dev` (notice
  the `gl-mac/` token);

PiperOrigin-RevId: 332094253