Add play-services-base to Android deps for analytics/admob/remote_config.
Due to a change in the transitive dependencies of analytics, play-services-base is no longer present. This is needed to check for the presence of Google Play Services. Admob and Remote Config both depend on analytics (through 'app'), which is how they previously got that dependency. Other APIs depend directly on play-services-base at the Android level.

PiperOrigin-RevId: 332267491