Handle permission errors when deleting intermediate directories
When running on Windows, an error occured due to a file in an intermediate directory still being open in another process when trying to delete the directory. The details are not clear as to why: my best guess is that the process opened by gradle still had one of the files it generated open, suggesting a (garbage collection?) race of some sort.

Since this occurs inconsistently, we can handle the error and raise a warning.