Workaround Windows Cloud SDK bug on GHA
The Google Cloud SDK's command line tools have an issue on Windows.
https://github.com/GoogleCloudPlatform/github-actions/issues/100

This causes steps using the command line tools to fail. This adds a workaround, which involves setting an environment variable. This was implemented in a way that will make it easy to revert once the issue is fixed.