Fix test lab paths on Windows and add checks
On GHA the the tool is complaining of a file not found, though it's not clear which one, only that it's happening during the gcs authorization check.

This puts paths through abspath and expanduser to ensure there aren't any issues with how they're formatted. It also adds check to ensure the Google Cloud command line tools are present, and the key file exists. This will make the cause of issues more obvious.