Fix a deadlock between snapshot listeners and cleanup.
The deadlock may be triggered by the following steps:
1. Add a snapshot listener to a collection (a document reference or a query will likely work as well, but not tested).
2. Wait for the listener to fire.
3. Immediately after that, delete the App or the Firestore instance (which invokes cleanup).

Note that because the deadlock is timing-based, it's not guaranteed to happen.

With simplifications, this leads to the following call stacks (using collection listeners for the sake of the example):

  User callback thread
- `AsyncEventListener::OnEvent` locks the listener mutex (LM);
- `AsyncEventListener::OnEvent` invokes the user callback;
- when the user callback is finished, a `QuerySnapshot` passed to the callback is destroyed
- the destructor of `QuerySnapshot` tries to unregister itself from the cleanup;
- the unregister function tries to acquire the cleanup mutex (CM).

  “Main” thread (the thread on which the Firestore instance is destroyed)
- Firebase App is destroyed;
- Firestore instance is destroyed;
- `Firestore` destructor triggers cleanup;
- Cleanup locks the cleanup mutex (CM);
- Cleanup invokes `firestore::ListenerRegistration::Cleanup`;
- Listener registration cleanup invokes `AsyncEventListener::Mute`;
- `AsyncEventListener::Mute` tries to acquire the listener mutex (LM).

To work around this, make sure to clear the listeners _before_ invoking the cleanup in `Firestore` destructor.

An alternative fix would be to make sure `CleanupNotifier` doesn't hold the mutex while invoking cleanup callbacks. Unfortunately, cleanup callbacks might destroy objects that are subject to cleanup as well and thus would try to unregister themselves, preventing this approach. In other words, if `CleanupNotifier` contains, say, 10 callbacks, there is no guarantee that it will still contain 10 callbacks after the first one is invoked.

PiperOrigin-RevId: 336603266