Added helper function to create directories recursively on desktop platforms(mac, windows and linux).
This fixes SEGFAULTS that were seen on running database integration tests. Issue was that "app_name" is usually a string with a "/" in it and mkdir calls on Mac and Windows were not creating directories recursively.

Changes were first done on github in this PR,
https://github.com/firebase/firebase-cpp-sdk/pull/124

Updated C++ and Unity release notes.

Created an accompanying unit test in `database/util_desktop_test UtilDesktopTest.TestSplitString`

Modified UtilDesktopTest.TestGetAppDataPath to test with an app path with a delimiter (the reason our previous release didn't catch this bug).

PiperOrigin-RevId: 337121244