Use ndk 16b
Run the "install prereqs" script for Android builds to install the NDK, then set env vars to use it in the build step. This isn't ideal since the setup script does more than just installing the NDK, but the additional work is minimal and low-risk. Re-using it also ensures the different workflows use the same logic to install the NDK. and put it in the same place.