Updated the raw_data field in messaging.h to be a vector of bytes instead of a string.
Using a string makes it difficult to work with byte streams that have embedded
null characters. In these situations it makes more sense to represent the data
as what it actually is, a sequence of raw bytes.

Previously this field was unused in C++ and never populated, so although this
is a change to a public header it should not break anyone, and anyone who is
broken by this was not using it correctly anyway.

Patch of CL 288600981, 288601322, 288601437

PiperOrigin-RevId: 339735518