Try fixing Android env vars on Windows integration tests
install_prereqs.sh is a bash script that downloads the ndk to /tmp/..., which gets translated to a different path on Windows. Setting the env var in the workflow file did not appear to cause a similar transformation. So instead we export the env vars using bash to mimick the effect.