Automated g4 rollback of changelist 335476967.
*** Reason for rollback ***

Firebase C++ now depends on Android M82, so this is fine to resubmit.

*** Original change description ***

Automated g4 rollback of changelist 333356092.

*** Reason for rollback ***

Temporary rollback to unblock the FPL release (which depends on Android M79).

*** Original change description ***

Set the language in Cloud headers to `gl-cpp/` in the C++ SDK on Android.

This is possible now that the Copybara CL cr/333174406 is submitted. This CL requires on Android functionality available starting from M80 (Firestore 21.7.0).

***

***

PiperOrigin-RevId: 340267852