#cleanup Unable to fix after @UiThread or runOnMainSync
(Not submitting until 11/9)

A go/rosie is not being done for this as adding the new calls and removing the old calls must be done at the same time (if we did one before the other, the tests would be failing in between)

For several years now, Android instrumentation tests in Google3 have been using the incorrect flag to specify test timeouts to AndroidJUnitRunner. However, due to an old bit of hacky behavior in Google3InstrumentationTestRunner, we cannot simply fix the problem by switching the code to use the correct flag.

In particular, Google3InstrumentationTestRunner currently starts a Looper that allows tests to be able to post events from the test thread (when tests aren’t supposed to be able to). Since using the correct flag in AndroidJUnitRunner would cause the test to now run in a separate thread from before, any tests that relied on this Looper would start to fail.

So, before we can turn on the flag, we need to go through and address all the tests that relied on the above behavior in Google3InstrumentationTestRunner.

If your test is in this changelist, it means that the test is posting events from the test thread, and we were unable to find a simple fix to suggest. To unblock everyone else, we’ve decided to move the associated Looper call to your tests’ setup so that your tests will still work when timeout enforcement is done correctly. A TODO(b/…) is being added next to the new call as a tracking bug for teams who wish to later (on their own time/schedule) refactor their test to no longer rely on the call.
```
# Startblock config for deferred review.
Startblock:
  has lgtm
  and then
  add tedbisson
```
PiperOrigin-RevId: 341435027