Fix Unity Editor hang on Windows when Listen() is invoked (b/172566004).
The issue is that when the C++ callback crosses the boundary into C# via P/Invoke, C# becomes aware of the existence of the thread. Later, when the "play" button is used in the Unity Editor to restart the app, the AppDomain unloading process starts, first signaling all threads that it knows exist then waiting for all of those threads to terminate; however, the callback thread is part of a thread pool and does not terminate in response to the signal, causing AppDomain unloading to deadlock.

This fix converts the callbacks to use `callback.h`, which was created a few years ago by the FPL team to address this exact issue. It addresses the issue because the thread that crosses the P/Invoke boundary is one that *is* terminated in response to the signal sent during AppDomain unloading, thus avoiding deadlock.

This fix probably also fixes b/170201731, where the Unity Editor hangs intermittently on non-Windows hosts as well.

PiperOrigin-RevId: 342100274