Re-write the transaction logic to run the user-supplied callback asynchronously.
The motivation for this change is to work around the `ThreadAbortException` that crashes the Unity Editor if Play/Pause is toggled while a transaction is in progress (b/169264353).

Previously, the user-supplied C# transaction callback was called from C++ and would block in C# until the callback, and all of its asynchronous tasks, especially I/O-bound `Transaction.GetSnapshotAsync()` tasks, completed. If the Play/Pause button was used to restart the app in the Unity Editor during the time that the C++ thread was blocked in C# then the `ThreadAbortException` would occur and unavoidably cross the P/Invoke boundary into C++, causing the crash.

The fix is to still call the user-supplied C# transaction callback synchronously from C++; however, instead of blocking in C# until all of the asynchronous tasks complete, the callback returns to C++ and C++ waits for C# to signal the completion of the asynchronous tasks. This reduces the window during which the `ThreadAbortException` could occur to be on par with other listener callbacks that do not suffer from `ThreadAbortException`-related crashes.

PiperOrigin-RevId: 342923297