Clean up `Firestore` instance management.
The main change is to merge the `firestore()`, `CachedFirestore()`, and `CreateFirestore()` methods into a single new method: `TestFirestore()`. The new `TestFirestore()` method accepts an app name that defaults to the default app's name.

The old methods map to the new methods as follows:
- `firestore()` -> `TestFirestore()`.
- `CachedFirestore(app_name)` -> `TestFirestore(app_name)`
- `CreateFirestore(app_name)` -> `db = TestFirestore(app_name)` followed by `DeleteFirestore(db)` or `RemoveFirestore(db)`

Finally, instead of explicitly calling `delete` on `App` and `Firestore` instances, use the methods `DeleteFirestore()` and `DeleteApp()`, respectively. All objects returned from `TestFirestore()` will be automatically deleted at the end of the test, so deleting these objects explicitly would result in double-deletion in the `FirestoreIntegrationTest` destructor. This is different from the old `CreateFirestore` method, which required that the caller explicitly delete the returned `Firestore` instance and its `App`.

PiperOrigin-RevId: 343482286