Improve integration test logging
These makes a couple of changes, all with the end goal of making it easier to debug failures:

1) Separate the building of desktop integration tests and their execution into separate steps. It was unintuitive that the build step ran desktop tests, but not mobile tests. A new script desktop_tester.py finds and executes desktop tests. Separating them into steps also organizes the logs better.

2) Rearrange log output for tests. Previously tests output logs in arbitrary order: now successes will report first, failures last, so that when scrolling up from the overall summary, one will immediately see the failures.

3) Add a workflow step to report terse summaries. This will make it possible to quickly determine what went wrong without trudging through multiple steps and long logs. It may still be necessary to go through detailed logs for more context, but this should help with determining if that's necessary.

To achieve this, every build/test script will report a summary, which gets written to a local file (summary.log in the top level output directory where the tests are built). The final step logs the contents of this file.

The new function in test_validation.py, 'summarize_tests_results' standardizes how all test runners output results.