Fix memory leaks in ReleaseXArrayElements:
The inspected code used JNI_COMMIT as a release mode,
which is specified as a release mode that copies
the changes to the Java array, but does not free the
native buffer (in case when the array is a copy) [1, 4].
Hence, when used as the last Release... mode, it effectively
leaks the native buffer when the provided array is a copy
(see the Hotspot implementation [2, 3], which always copies).

Instead, in this use-case "0" mode must be used.

The code fragments have been inspected to not do anything fancy with
the buffer pointer — they all store the buffer in local variables,
and release the array after they are done.

A full-blown fix is to use a safer alternative, jni_helper,
as `CHECK_JNI(FATAL, env).GetXArrayElements`, see
http://google3/util/java/jni_helper.h?l=141-150&rcl=341324486

Ironically, the issue was discovered when I searched for any good examples
of JNI_COMMIT usage IRL, to aid in making a decision on supporting this
flag in jni-rs, safe JNI bindings for Rust.

The code under third_party is not yet inspected.

[1]: https://docs.oracle.com/en/java/javase/11/docs/specs/jni/functions.html#releaseprimitivetypearrayelements-routines
[2]: http://google3/third_party/java_src/jdk/head/src/src/hotspot/share/prims/jni.cpp?l=2613&rcl=302688582
[3]: http://google3/third_party/java_src/jdk/head/src/src/hotspot/share/prims/jni.cpp?l=2665-2670&rcl=302688582
[4]: https://developer.android.com/training/articles/perf-jni#primitive-arrays

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:341596811:BASE:341578570:1605013504633:7ec031b2
PiperOrigin-RevId: 345402895