Fix bug where including a "delete" operation in an update causes partial updates.
The easiest fix would have been to replace `break` with `continue`; however, I changed it to an if/else statement to be in line with the other SDKs. See b/174676322 for details.

This bug affects both the Firestore Unity and C++ SDKs. Therefore, I have added unit test coverage to both for this specific case.

This bug was originally reported by a customer: https://github.com/firebase/quickstart-unity/issues/882

PiperOrigin-RevId: 345664484