Delete unused (and invalid) overload of Promise::RegisterForTask().
The method is "invalid" because, when instantiated, it fails to compile (the `RegisterForTask` overload that it calls does not exist). This was never noticed because this type of compilation error is not detected until the method is actually invoked, forcing the compiler to instantiate it. Since the method was never invoked, nobody noticed.

PiperOrigin-RevId: 347404402