Run gtest download script instead of relying on cmake
Currently, we rely on some cmake logic to setup gtest, but this is flaky. To work around this, we run the gtest download script before running the integration test setup script: the latter will copy over gtest alongside the test framework files into the project. This will result in the gtest cmake logic being bypassed.

With this in place, the previous workaround of running the android build twice in the presence of a specific failure is no longer necessary. It has been removed.