Support both Linux ABIs, legacy and C++11 (#213)
This change keeps the legacy ABI but also adds an option to build using the C++11 ABI if FIREBASE_LINUX_USE_CXX11_ABI is set during the cmake configure step.

In packaging, the location of the Linux libraries is now libs/linux/$arch/$abi, where $arch is x86_64 or i386 (as before), and $abi is legacy or cxx11. For example, libs/linux/x86_64/cxx11/libfirebase_app.a. The binary package's CMakeLists.txt has been updated to choose the correct one based on whether the compiler option is enabled (it checks the actual CFLAGS for the option, not the Firebase-specific flag, so it should handle the selection automatically even if the developer's CMake project doesn't provide the Firebase-specific flag).

Also any sub-builds (e.g. boringssl) will also be configured with the requested ABI option.