Change desktop packaging script to run merge_libraries in parallel. (#218)
By far the slowest part of C++ desktop packaging is running merge_libraries, especially on Windows. This PR allows the merge_libraries step -- which uses Linux runners to merge Windows libraries - to run in parallel as long as the "parallel" command is installed on the machine, if you pass in the -j flag.

To enable this change, a bit of code is added to merge_libraries.py to lock and unlock the cache file (which is used across merge_libraries instances) using flock(). It requests a shared lock while reading, and an exclusive lock while writing. The flock() call will block until the process can obtain the lock.

This change saves nearly an hour in packaging time for the longest packaging platform, Windows Debug X64 Static (and Dynamic).