Using the recommended backwards compatible way of scheduling jobs introduced in Android O.
We moved from using `IntentService` to `JobIntentService` in preparation for this change but the call to schedule jobs remained the same which meant that our custom IntentService was not doing any work at all.

Using `JobIntentService.enqueueWork`, when running on Android O or later, the work will be dispatched as a job via JobScheduler.enqueue. When running on older versions of the platform, it will use Context.startService.
https://developer.android.com/reference/androidx/core/app/JobIntentService

Fixes
https://github.com/firebase/quickstart-unity/issues/877
https://github.com/firebase/quickstart-unity/issues/883

PiperOrigin-RevId: 350176964