Add JNI wrapper for the GMS `Task` class and modify the Firestore Android C++ SDK implementation to use it.
This change has two benefits: It improves readability of the code because variables will be defined as `Local<Task>` instead of `Local<Object>` and also improves type safety because it is no longer possible to specify any object when a `Task` is specifically required.

The motivation for this change also includes future unit tests which will interact with `Task` objects directly, instead of encapsulating them in `Promise` objects.

As part of this change, I also created a new `FirestoreAndroidIntegrationTest` class, which extends `FirestoreIntegrationTest` and adds Android-specific functionality. I modified the existing `PromiseTest` to use it, and the new `TaskTest` that was added uses it as well.

PiperOrigin-RevId: 351816780