Use the Objective-C CredentialsProvider on iOS
Previously, we used the C++ CredentialsProvider on all non-Android platforms.
This works because Firebase Auth's C++ interface is portable.

While switching to use Auth via the FunctionRegistry mechanism (to make Auth a
soft dependency of Firestore), it's become clear that the FunctionRegistry
mechanism is only available on desktop platforms. This means we can't use C++
Auth on iOS.

PiperOrigin-RevId: 353057413