Allow packaging step to specify an explicit binutils format, and add some configuration options to build_desktop (#241)
Specifying an explicit binutils format fixes an issue in Windows builds where some libraries would be autodetected, but then would create invalid output libraries since the target format wasn't specified.

This was causing a linker issue with absl in Firestore with vector deleting destructors not being found, because certain absl libraries were not created valid by merge_libraries. In particular, Windows libraries are now set to be output in pe-bigobj-x86-64/pe-bigobj-i386 format, since Firestore in particular has >65K sections and requires bigobj. There's no harm in making all the other libraries the same format.

This PR also adds a couple of flags to build_desktop.py:

It adds support for performing a verbose build during packaging, in case it's needed for further debugging.
Adds a flag for disabling vcpkg, although we don't yet use it as we require vcpkg to get the "protoc" executable.
It also includes a bit of code in the workflow to handle Mac arm64 although we don't yet build for that platform.