Fixed issue with integration tests finding google-services.json
Changed the search paths for the google-services.json file in
CMakeLists.txt files to use full paths instead of relative paths.

According to the CMake documentation, the `if (EXISTS ...)` condition is
only well defined for full paths. This might work on some platforms but
it failed in my tests on Windows.