Port exception-throwing abstractions from iOS to common
Unfortunately, because we still support STLPort on Android, this requires
special casing there. On desktop platforms this delegates back to the iOS core.

This includes a change to the iOS firestore_exceptions.h which needs to be
ported to GitHub before the next import. It seems as if policy checks in Abseil
fail on Android even when compiling against non-STLPort C++ runtimes.

PiperOrigin-RevId: 354562822