Expand validation_test to cover other platforms
Also fix transactions on Android to throw immediately when C++
exceptions are enabled. While this doesn't have much of an effect on
C++ users, this will make it possible for the Unity client to
synchronously throw in these cases, which is expected in C#.

Cloned from CL 355019084 by 'g4 patch'.
Original change by mcg@mcg:fig-export-firestore-change-286-4f589224b545:4324:citc on 2021/02/01 13:52:36.

PiperOrigin-RevId: 359824564