Fixed a crash that can occur as a result of a race condition when rapidly adding and removing a listener.
When a listener is added or removed, it doesn't get added or removed immediately. Instead, a transaction is scheduled to run at some point in the future. If the developer adds and removes a listener in quick succession and then deletes it, those scheduled callbacks might not run until after the listener has already been deleted, resulting in calling a member function of a now garbage pointer and crashing.

The solution is to have each ValueListener and ChildListener keep track of all locations where they are being used. If the listener is deleted or removed, then it invalidates references to it in a mutex protected manner. Then, when the background thread's scheduled callback gets around to adding it, it will see that the listener should not be added and early exit.

Because of the mutex protection on each Listener, this broke SingleValueListeners which are used in the GetValue functions for retrieving a single value, since they cannot destroy a mutex they currently have locked. Instead, we schedule a callback to do the cleanup.

PiperOrigin-RevId: 360516349