Fixed race condition when removing listeners.
Fixed a crash that can occur as a result of a race condition when
rapidly removing and deleting a listener.

When a listener is added or removed, it doesn't get added or removed
immediately. Instead, a transaction is scheduled to run at some point in
the future. If the developer adds and removes a listener in quick
succession and then deletes it, those scheduled callbacks might not run
until after the listener has already been deleted, resulting in calling
a member function of a now garbage pointer and crashing.

The solution is to immediately mark a listener as having been removed
from the main thread (as well as scheduling it to be removed from the
background thread, as it was already doing). Any new event that come in
on a listener that has been marked removed will be ignored. The
developer is free to delete their Listener after that point without fear
that an event fire from the background thread after it has been deleted.