Deflake: Cache NDK Downloads (#327)
There have been many workflow runs which have failed to download the NDK. I'mt not sure if this is due to its size or if the GHA Vms have a bad connection, etc, but more often than not when theres a failed download, it's the NDK.

Adding steps to cache the NDK to the Integration, Packaging and Android workflows. The NDK cache is based on the host OS and is based on the path of the android/install_prereqs.py script.

Includes some indentation changes to the Android Prereqs script, but no logical changes.