Trigger integration tests by setting a "tests-requested" label. (#339)
This PR adds some label-based triggers for integration tests, which allow the integration tests to be requested on a PR by setting the "tests-requested" label.

There are two modes for requesting tests, "quick" and "full":

"quick" tries to be clever and figure out a subset of APIs (and even platforms) to test. This heuristic is really simple right now (it just looks at directory names and, for platform, filenames) and if there is any ambiguity it will just run the normal suite of tests.

"full" just runs all the tests. Currently this is set to use an expanded test matrix, though this could be changed in the future.

The test workflow handles updating the labels and posting/updating a comment with the status of the integration tests as they run. All of the jobs that mess with labels and comments are skipped if the workflow wasn't run from the PR label.

(See #337—which won't be merged—for an example of "quick" mode in action, autodetecting to just run a subset of tests.)