Added a new test suite for SyncPoints.
The new test suit is driven by the same master set of test cases that
drive SyncPoint tests on the iOS and Android implementation.

All tests are run twice, once at the root, and once again at a deeper
level, to ensure that the depth of the changes made does not matter.

Some tests currently fail. It's not clear if these are legitimate bugs
in the test suite, or if there are some bugs in the DB implementation.
For the time being, these tests are skipped using `GTEST_SKIP();` and
are annotated with whether they result in an assert being triggered or
just a failure due to unmet expectations.

Also included is a minor change to what is considered a valid priority
(null priorities are valid) and how QueryParams determine if they have a
start or end ranges (null values are allowed, if there is a start or end
key name supplied). I tried to keep bug fixes in separate commits, but
allowing null was necessary or almost none of the tests would pass.