Made parameters in QueryParams optional.
QueryParams previous used null variants and empty strings to indicate
that a field should not be used as a filter. This creates a problem when
you want to filter by something that has a null value or an empty
string. To fix this std::string and Variant have been replaced with
Optional<std::string> and Optional<Variant> so that "filter on null" is
distinct from "there is no filter".