Made changes to QueryParams more narrow.
Previously the changes to QueryParams so that they included optional
fields had cascaded out to a number of other classes and functoins, but
the bug being fixed didn't require the changes to extend to so many
places. This narrows the scope of the changes to just the necessary
classes and functions.