boringSSL clone specific commit (#485)
Clone the boringSSL sources to a specific commit, instead of cloning the repo first and then checking out a commit.

This frees up enough disk space that we no longer need to copy the resulting artifacts over to the C drive on Windows hosts in order to make enough room for the google cloud sdk installation. Therefore I've removed that step from the CI saving up to 20 minutes per windows integration test build.