Run lint checks on PRs, posting any warnings as comments directly on the files (#508)
This script performs a number of steps:
1. Get the PR's diff to find the list of affected files and lines in the PR.
2. Run cpplint on all files in the PR. For each lint warning, check if it falls within the range of lines affected by the diff. Omit that warning if it doesn't fall in the affected lines.
3. Delete any prior lint warning comments posted by previous runs (this takes a moment as they must be deleted one by one, due to GitHub API limitations)
4. Post all the lint warnings that fall within the range of the PR's diff (this is done all at once), on the PR's most recent commit.

It will ignore any lint warnings that occur outside the PR's diff, and it will only run cpplint on files that are added/modified in the PR in the first place. There is also an exclusion list to exclude sets of files (via regex).

Any other errors from cpplint (e.g. "I don't know how to parse this file") are simply ignored.

Also, the whole process will be skipped if the PR has the 'no-lint' label.