Use git apply patch to fix uWebsockets crash, previously fixed by python script in #344 (#507)
Previously we downloaded a tarball release of uWebSockets and then patched the code to fix a crash issue. The patch was executed by a custom python script which somewhat blindly overwrote at set of line numbers in the file. See PR #344

Git apply patch wasn't used because it requires a initialized git repo to work. Cloning the the uWebsockets repo with history and then checking out to an older commit would explode the disk usage, causing parition exhaustion issues on Windows, so we were forced to use the non-git-initialized tarball.

Thankfully I found a way to fetch a repo at a specific commit, removing the overhead of cloning all of the history. This change updates the cmake download step for uWebSockets to fetch the commit instead of downloading a tarball. With a git initialized repo we can our change with a far safer git apply patch instead of our custom python script, akin to what was done for boringSSL in PR #485