Standardize header guards (#518)
Standardize header guards, and turn on lint warnings for bad header guard names and line length, which were previously disabled. Includes a small fix to replace the double-quote symbol (") with the html entity (ampersand-quot;) in the lint output, for HTML compatibility.

This also turns off the "build/include_subdir" and "whitespace/indent" lint warnings as they are extraneous.