On PRs from external users, dismiss stale reviews when a new commit is pushed. (#572)
With our new project settings, when a PR is approved, that approval will stay valid even if further commits are pushed into that PR's branch.

For external users, though, we want to invalidate those approvals if any additional commits are made subsequent to the original approval. In this case, we consider anyone without "admin" access to be an "external user", and any PR from a forked repo to be an "external PR". In either of those scenarios, stale approvals will be dismissed.

This PR adds a workflow that runs on pull_request_target, which runs inside the `main` branch and has enough access to dismiss reviews. It also adds a script that allows you to dismiss reviews on a PR.